+++ src/tests/slider/Slider.svelte
@@ -5,7 +5,8 @@
 	export let min = 0;
 	export let step = 1;
 	const {
-		elements: { root, range, thumb },
+		elements: { root, range, thumb, tick },
+		states: { ticks }
 	} = createSlider({
 		defaultValue: value,
 		max,
@@ -26,4 +27,7 @@
 			class="block h-5 w-5 rounded-full bg-white focus:ring-4 focus:ring-black/40"
 		/>
 	</span>
+	{#each { length: $ticks } as _}
+		<span use:melt={$tick()} data-testid="tick" />
+	{/each}
 </main>